Move `get_stamp_size` to `pipeline`
